# docs/swagger.yaml
definitions:
...
  handlers.LoginRequest:
    properties:
      identifier:
        description: 可以是用户名或邮箱
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
...
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 255
        minLength: 6
        type: string
      username:
        maxLength: 64
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  xerr.Response:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 响应数据
      message:
        description: 消息
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口
      parameters:
      - description: 登录信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回token
          schema:
            $ref: '#/definitions/xerr.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/xerr.Response'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/xerr.Response'
      summary: 用户登录
      tags:
      - 用户认证
...
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: 用户注册接口
      parameters:
      - description: 注册信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/xerr.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/xerr.Response'
        "409":
          description: 用户名或邮箱已存在
          schema:
            $ref: '#/definitions/xerr.Response'
      summary: 用户注册
      tags:
      - 用户认证
...
swagger: "2.0"